+++ hc/accounts/models.py
@@ -60,6 +60,7 @@ class Profile(models.Model):
     team_limit = models.IntegerField(default=2)
     sort = models.CharField(max_length=20, default="created")
     deletion_notice_date = models.DateTimeField(null=True, blank=True)
+    last_active_date = models.DateTimeField(null=True, blank=True)
 
     objects = ProfileManager()